Read serial ouput from EZ0
This will use the serial output from the EZ0 instead of
the pulse output.
The distance is output in whole inches, so the code
has been refactored to handle this.

This removes much of the noise we have seen in the distance
readings, as the noise were introduced when we were reading the pulses.